app/requirements.txt
Flask
boto3
awsiotsdk